+++ packages/node/src/subcommands/testing.module.ts
@@ -8,6 +8,7 @@ import {
   ApiService,
   ConnectionPoolService,
   DbModule,
+  NodeConfig,
   PoiService,
   StoreService,
   TestRunner,
@@ -49,16 +50,23 @@ import { TestingService } from './testing.service';
         project: SubqueryProject,
         connectionPoolService: ConnectionPoolService<EthereumApiConnection>,
         eventEmitter: EventEmitter2,
+        nodeConfig: NodeConfig,
       ) => {
         const apiService = new EthereumApiService(
           project,
           connectionPoolService,
           eventEmitter,
+          nodeConfig,
         );
         await apiService.init();
         return apiService;
       },
-      inject: ['ISubqueryProject', ConnectionPoolService, EventEmitter2],
+      inject: [
+        'ISubqueryProject',
+        ConnectionPoolService,
+        EventEmitter2,
+        NodeConfig,
+      ],
     },
     SchedulerRegistry,
     TestRunner,