+++ packages/node/CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Synced with main sdk:
+  - add `--primary-network-endpoint` cli option
+  - Support for update availability logs
+
 ## [2.10.0] - 2023-07-31
 ### Added
 - Added `!null` filter for logs (#135)