+++ packages/types/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- fix missing `unsafeApi` in the global
+
 ## [2.2.3] - 2023-07-31
 ### Fixed
 - Update license (#137)